+++ examples/app/SimplePattern/index.tsx
@@ -18,7 +18,7 @@ function SimplePattern() {
         <Item img="https://unsplash.it/479/205" />
       </Carousel>
       <h6>
-        <a href="https://github.com/FormidableLabs/react-swipeable/blob/main/examples/app/SimplePattern/pattern.tsx">See code</a> for example usage of <code style={{ whiteSpace: "nowrap" }}>touch-action: none</code> to prevent scrolling.
+        <a href="https://github.com/FormidableLabs/react-swipeable/blob/main/examples/app/SimplePattern/pattern.tsx">See code</a> for example usage of <code>onTouchStartOrMouseDown</code> and <code>touchEventOptions</code>
       </h6>
     </div>
   );